Added 2-bit branch predictor
Added 2-bit branch predictor that outputs the default decision on branches, and an override signal if the previous branch decision was wrong. The MSB of the branch predictor is the default decision, and the override signal is set when the default decision does not match the zero signal output from the branch calculations in the ID stage.